refactor: move WatchHandle into reactivity
